RecyclerView with Data Binding using Network Call

+++ app/src/main/java/com/example/weatherreport/models/Wind.java
@@ -6,13 +6,13 @@ public class Wind {
     @SerializedName("speed")
     private Double speed;
     @SerializedName("deg")
-    private Integer deg;
+    private Double deg;
 
     public void setSpeed(Double speed) {
         this.speed = speed;
     }
 
-    public void setDeg(Integer deg) {
+    public void setDeg(Double deg) {
         this.deg = deg;
     }
 
@@ -20,7 +20,7 @@ public class Wind {
         return speed;
     }
 
-    public Integer getDeg() {
+    public Double getDeg() {
         return deg;
     }
